Split resolver body local tests
