+++ src/typechecker/tests/resolver_locals.rs
@@ -1,6 +1,7 @@
 use super::*;
 
 mod absence;
+mod body_locals;
 
 #[test]
 fn check_program_with_symbols_requires_resolver_parameter_locals() {
@@ -210,185 +211,5 @@ main = () i32 {
             .message
             .contains("resolver local symbol 'value' has mutability mutable, expected immutable")),
         "expected scoped resolver local mutability diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_requires_resolver_pattern_locals() {
-    let program = parse_program(
-        r#"
-Option:
-    None,
-    Some(i32)
-
-main = (value: Option) i32 {
-    value ?
-        | Some(inner) { inner }
-        | None { 0 }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.remove_for_test(Namespace::Local, "inner");
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("missing resolver pattern local should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver symbol table missing local symbol 'inner'")),
-        "expected missing resolver pattern local diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_requires_resolver_top_level_expr_locals() {
-    let program = parse_program(
-        r#"
-value := 1
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.remove_for_test(Namespace::Local, "value");
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("missing resolver top-level expr local should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver symbol table missing local symbol 'value'")),
-        "expected missing resolver top-level expr local diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_requires_resolver_closure_locals() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    mapper = (input: i32) i32 {
-        inner = input
-        inner
-    }
-    0
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.remove_for_test(Namespace::Local, "inner");
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("missing resolver closure local should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver symbol table missing local symbol 'inner'")),
-        "expected missing resolver closure local diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_validates_resolver_closure_parameter_mutability() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    mapper = (mut input: i32) i32 {
-        input = input + 1
-        input
-    }
-    0
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_local_mutability_for_test("input", Some(false));
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("resolver closure parameter mutability mismatch should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver local symbol 'input' has mutability immutable, expected mutable")),
-        "expected resolver closure parameter mutability diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_requires_resolver_struct_field_default_locals() {
-    let program = parse_program(
-        r#"
-Point: {
-    x: i32 = {
-        value = 1
-        value
-    }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.remove_for_test(Namespace::Local, "value");
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("missing resolver struct field default local should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver symbol table missing local symbol 'value'")),
-        "expected missing resolver struct field default local diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_requires_resolver_behavior_default_locals() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    to_json: (Self) str {
-        value = "{}"
-        value
-    }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.remove_for_test(Namespace::Local, "value");
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("missing resolver behavior default local should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver symbol table missing local symbol 'value'")),
-        "expected missing resolver behavior default local diagnostic, got {err:?}"
     );
 }
